CRUD with auth done

+++ controller/feed.js
@@ -6,6 +6,7 @@ const { validationResult } = require('express-validator/check')
 
 // local imports
 const Post = require('../models/post')
+const User = require('../models/user')
 
 
 
@@ -44,21 +45,29 @@ exports.createPost = ((req, res, next) => {
     const title = req.body.title
     const content = req.body.content
     const imageUrl = req.file.path
+    let creator
 
     const post = new Post({
         title: title,
         content: content,
         imageUrl: imageUrl,
-        creator: {
-            name: 'Ali Raza 2'
-        },
+        creator: req.userId
     })
     post.save().then(result => {
+        return User.findById(req.userId)
+    }).then(user => {
+        creator = user
+        user.posts.push(post)
+        return user.save()
+    })
+    .then(result =>{
         res.status(201).json({
             message: 'Post Created Successflly',
-            posts: result
+            posts: post,
+            creator: {_id: creator._id, name: creator.name }
         })
-    }).catch(err => {
+    })
+    .catch(err => {
         if (!err.statusCode) {
             err.statusCode = 500
         }
@@ -109,6 +118,11 @@ exports.updatePost = (req, res, next) => {
             error.statusCode = 404
             throw error
         }
+        if(post.creator.toString() !== req.userId){
+            const error = new Error('Not Authorized!')
+            error.statusCode = 403
+            throw error
+        }
         if (imageUrl !== post.imageUrl) {
             clearImage(post.imageUrl)
         }
@@ -137,17 +151,29 @@ exports.deletePost = (req, res, next) => {
                 throw error
             }
         }
+        if(post.creator.toString() !== req.userId){
+            const error = new Error('Not Authorized!')
+            error.statusCode = 403
+            throw error
+        }
         clearImage(post.imageUrl)
         return Post.findByIdAndRemove(postId)
     }).then(result => {
+        return Post.findById(req.userId)
+    })
+    .then(user =>{
+        user.posts.pull(postId)
+        return user.save()
+    })
+    .then(result =>{
         res.status(200).json({ message: 'Deleted Post' })
     })
-        .catch(err => {
-            if (!err.statusCode) {
-                err.statusCode = 500
-            }
-            next(err)
-        })
+    .catch(err => {
+        if (!err.statusCode) {
+            err.statusCode = 500
+        }
+        next(err)
+    })
 }
 
 const clearImage = filePath => {
